field validation container
