+++ components/shared/Form/Field/styles.js
@@ -1,7 +1,15 @@
 import {StyleSheet} from 'react-native'
 
+import * as colors from '@/assets/colors'
+
 export default StyleSheet.create({
-  child: {
+  container: {
     marginBottom: 15
+  },
+  text: {
+    fontSize: 12,
+    fontWeight: '500',
+    color: colors.red.medium,
+    marginTop: 5
   }
 })
