+++ components/shared/Form/fields/TextInput.js
@@ -1,5 +1,5 @@
 import TextInput from '@/components/shared/TextInput'
-import {field} from '../Context/Consumer'
+import {field} from '../Field'
 
 export default field()(({onChange, ...props}) => (
   <TextInput {...props} onChangeText={onChange} />
